Update code to ES6 standards & work with the latest Chart.js
